support web multi-browser parallel

+++ flybirds/report/fail_feature_create.py
@@ -2,14 +2,21 @@
 """
 fail scenario create rerun
 """
+import base64
 import json
 import os
+import random
 import re
 import shutil
-import subprocess
+from concurrent.futures.thread import ThreadPoolExecutor
+from functools import partial
+from multiprocessing import Pool
+from subprocess import Popen
 
 from flybirds.core.config_manage import FlowBehave
 from flybirds.report import json_format_deal
+from flybirds.report.parallel_runner import get_features_num, \
+    execute_parallel_feature
 from flybirds.report.rerun_params import get_rerun_params
 from flybirds.utils import file_helper
 from flybirds.utils import flybirds_log as log
@@ -40,6 +47,7 @@ class FailScenarioSum:
         scenario_value = FailScenarioInfo(
             feature_name, scenario_name, description
         )
+        # fail_relevance message
         self.fail_scenarios[scenario_key] = json.dumps(scenario_value.__dict__)
         return scenario_key
 
@@ -70,10 +78,15 @@ class FailScenarioInfo:
             self.description = description
 
 
-def rerun_launch(need_rerun_args, report_dir_path, run_args):
+def rerun_launch(context, is_parallel):
     """
     start to rerun
     """
+    # get run args
+    run_args = context.get("run_args")
+    need_rerun_args = context.get("need_rerun_args")
+    report_dir_path = context.get("report_dir_path")
+
     # Determine whether the failed scenario needs to be re-run
     flow_behave_config = FlowBehave({}, None)
     rerun_report_dir_path = None
@@ -88,6 +101,7 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
         max_retry_count = flow_behave_config.max_retry_count
         run_count = 1
         rerun_dir_path = report_dir_path
+        # create files under rerun1 except for the 'result'
         count_satisfy = create_rerun(
             report_dir_path,
             rerun_dir_path,
@@ -96,13 +110,15 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
         )
         # The number of failures is not higher than the limited value and
         # needs to be re-run
+
+        # execute the rerun cmd, and copy the results to report dir
         if count_satisfy:
             while max_retry_count > 0:
                 log.info(
                     (
                         f"Number of failed retries{run_count}, start to "
                         f"generate feature files that failed to rerun, "
-                        f"report root directory{report_dir_path}"
+                        f"report root directory {report_dir_path}"
                     )
                 )
                 rerun_dir_path = report_dir_path
@@ -123,6 +139,7 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
                         f"{max_retry_count}, run_count: {run_count}"
                     )
                     if count_satisfy:
+                        # 1.get behave rerun_cmd_str,rerun_report_dir_path
                         rerun_params = get_rerun_params(
                             run_count,
                             rerun_feature_path,
@@ -135,17 +152,10 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
                             "rerun_report_dir_path"
                         )
                         if rerun_cmd_str is not None:
-                            # cwd_pth = os.getcwd()
-                            # if os.environ.get('base_dir') is not None:
-                            #     cwd_pth = os.environ.get('base_dir')
-                            rerun_behave_process = subprocess.Popen(
-                                rerun_cmd_str,
-                                cwd=os.getcwd(),
-                                shell=True,
-                                stdout=None,
-                            )
-                            rerun_behave_process.wait()
-                            rerun_behave_process.communicate()
+                            # 2. execute rerun_cmd_str
+                            failed_rerun(
+                                rerun_cmd_str, rerun_feature_path,
+                                context, is_parallel)
                             # Re-run of the failed scenario ends
                             # Number of reruns -1, actual number of runs +1
                             max_retry_count = max_retry_count - 1
@@ -156,6 +166,7 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
                                 f"{report_dir_path}, rerun_report_dir_path:"
                                 f" {rerun_report_dir_path}"
                             )
+                            # 3. copy the execution results to report dir
                             json_format_deal.copy_rerun_screen(
                                 report_dir_path, rerun_report_dir_path
                             )
@@ -179,12 +190,12 @@ def rerun_launch(need_rerun_args, report_dir_path, run_args):
                 f"Do you need to rerun the task: {count_satisfy}, the failed "
                 f"retry task was not executed"
             )
-        log.info(
-            f"Start processing the json report under the directory "
-            f"{report_dir_path}"
-        )
-    # if rerun_report_dir_path is not None:
-    json_format_deal.parse_json_data(report_dir_path, rerun_report_dir_path)
+    log.info(
+        f"Start processing the json report. report_dir_path: "
+        f"[{report_dir_path}],rerun_report_dir_path:[{rerun_report_dir_path}]"
+    )
+    json_format_deal.parse_json_data(report_dir_path, rerun_report_dir_path,
+                                     is_parallel)
 
 
 def create_rerun(report_dir, rerun_dir, run_count, max_fail_count=1.0):
@@ -199,8 +210,9 @@ def create_rerun(report_dir, rerun_dir, run_count, max_fail_count=1.0):
     fail_count = 0
 
     rerun_root_dir = f"{rerun_dir}{os.sep}rerun{run_count}"
-    if not file_helper.create_dirs(rerun_root_dir):
-        file_helper.clear_dirs(rerun_root_dir)
+    file_helper.create_dirs(rerun_root_dir)
+    # if not file_helper.create_dirs(rerun_root_dir):
+    #     file_helper.clear_dirs(rerun_root_dir)
 
     fail_scenario_static = FailScenarioSum()
 
@@ -209,7 +221,6 @@ def create_rerun(report_dir, rerun_dir, run_count, max_fail_count=1.0):
 
     # Second level file index
     # rerun_feature_second_index = 1
-
     sum_count, fail_count, fail_scenario_static, rerun_root_dir = \
         process_loop_block(report_dir, rerun_feature_index, sum_count,
                            fail_count,
@@ -240,8 +251,9 @@ def create_rerun(report_dir, rerun_dir, run_count, max_fail_count=1.0):
             fail_count > max_fail_count,
         )
         result = False
-
+    # 2. write fail_relevance.json file
     fail_scenario_static.serialize_to_file(rerun_root_dir)
+    # 3. copy environment.py and so on
     copy_behave_need_file(rerun_root_dir)
     return result
 
@@ -249,6 +261,11 @@ def create_rerun(report_dir, rerun_dir, run_count, max_fail_count=1.0):
 def process_loop_block(report_dir, rerun_feature_index, sum_count, fail_count,
                        exist_scenario_name, fail_scenario_static, run_count,
                        rerun_root_dir):
+    """
+    iterate through all json in the report_dir
+    1.find  the cases that need to be rerun and write them to the feature file
+    2.modify failed status to rerun status,and re-write json file
+    """
     for file_item in os.listdir(report_dir):
         if re.search(r"\.json", str(file_item)) is not None:
             # noinspection PyBroadException
@@ -256,6 +273,7 @@ def process_loop_block(report_dir, rerun_feature_index, sum_count, fail_count,
                 file_path = os.path.join(report_dir, file_item)
                 report_json = file_helper.get_json_from_file_path(file_path)
                 if isinstance(report_json, list):
+                    # 1. to find need to be rerun case
                     for feature in report_json:
                         cur_feature_array = get_init_feature_array(
                             rerun_feature_index, feature.get("language")
@@ -293,9 +311,9 @@ def process_loop_block(report_dir, rerun_feature_index, sum_count, fail_count,
                                     rerun_scenario_name = scenario["name"]
                                     while (
                                             rerun_scenario_name
-                                            in exist_scenario_name
-                                    ):
-                                        rerun_scenario_name += str(range(9))
+                                            in exist_scenario_name):
+                                        rerun_scenario_name += str(
+                                            random.randint(0, 10))
                                     exist_scenario_name.append(
                                         rerun_scenario_name
                                     )
@@ -358,6 +376,7 @@ def process_loop_block(report_dir, rerun_feature_index, sum_count, fail_count,
                                         cur_feature_array,
                                     )
 
+                    # 2.re-write json file
                     file_helper.store_json_to_file_path(
                         report_json, file_path, "w"
                     )
@@ -394,19 +413,25 @@ def copy_behave_need_file(rerun_root_dir):
     # elif os.environ.get('base_feature_dir') is not None:
     #     cwd_pth = os.environ.get('base_feature_dir')
     file_helper.create_dirs(os.path.join(rerun_root_dir, "steps"))
-    file_helper.clear_dirs(os.path.join(rerun_root_dir, "steps"))
-    shutil.copy(
-        os.path.join(os.getcwd(), "features", "steps", "steps.py"),
-        os.path.join(rerun_root_dir, "steps", "steps.py"),
-    )
-    shutil.copy(
-        os.path.join(os.getcwd(), "features", "environment.py"),
-        os.path.join(rerun_root_dir, "environment.py"),
-    )
-    shutil.copy(
-        os.path.join(os.getcwd(), "features", "__init__.py"),
-        os.path.join(rerun_root_dir, "__init__.py"),
-    )
+    # file_helper.clear_dirs(os.path.join(rerun_root_dir, "steps"))
+    steps_path = os.path.join(os.getcwd(), "features", "steps", "steps.py")
+    if os.path.exists(steps_path):
+        shutil.copy(
+            os.path.join(os.getcwd(), "features", "steps", "steps.py"),
+            os.path.join(rerun_root_dir, "steps", "steps.py"),
+        )
+    env_path = os.path.join(os.getcwd(), "features", "environment.py")
+    if os.path.exists(env_path):
+        shutil.copy(
+            os.path.join(os.getcwd(), "features", "environment.py"),
+            os.path.join(rerun_root_dir, "environment.py"),
+        )
+    init_path = os.path.join(os.getcwd(), "features", "__init__.py")
+    if os.path.exists(init_path):
+        shutil.copy(
+            os.path.join(os.getcwd(), "features", "__init__.py"),
+            os.path.join(rerun_root_dir, "__init__.py"),
+        )
 
 
 def set_rerun_info(user_data, gr):
@@ -445,3 +470,59 @@ def set_rerun_info(user_data, gr):
             gr.set_value("rerunFailInfo", last_fail_scenario_info_obj)
 
     log.info(f"user data，count:{len(user_data)}")
+
+
+def failed_rerun(rerun_cmd_str: str, rerun_feature_path, context,
+                 is_parallel):
+    if is_parallel:
+        parallel_rerun(rerun_cmd_str, rerun_feature_path, context)
+    else:
+        rerun_behave_process = Popen(
+            rerun_cmd_str,
+            cwd=os.getcwd(),
+            shell=True,
+            stdout=None,
+        )
+        rerun_behave_process.wait()
+        rerun_behave_process.communicate()
+
+
+def parallel_rerun(rerun_cmd_str: str, rerun_feature_path, context):
+    log.info('parallel_rerun start!')
+    dry_cmd = f'behave {rerun_feature_path} -d -k -f json --no-summary'
+    features = get_features_num(dry_cmd)
+    log.info(f'[parallel_rerun] features num: {len(features)}')
+
+    context['rerun_cmd_str'] = rerun_cmd_str
+    context['rerun_feature_path'] = rerun_feature_path
+
+    log.info('start rerun thread...')
+    with ThreadPoolExecutor(max_workers=3) as t_pool:
+        browser_types = context.get('browser_types')
+        for b_type in browser_types:
+            t_pool.submit(rerun_multiplication, b_type, context, features)
+    log.info('all rerun thread done...')
+
+
+def rerun_multiplication(browser_type, context, features):
+    log.info('rerun_multiplication start')
+    log.info(f'rerun_multiplication browser_type:{browser_type}')
+
+    rerun_cmd_str = context.get("rerun_cmd_str")
+    rerun_feature_path = context.get("rerun_feature_path")
+    processes = context.get("processes")
+
+    cur_browser_type = str(base64.b64encode(browser_type.encode('utf-8')),
+                           'utf-8')
+    rerun_cmd_str = rerun_cmd_str + f'  -D cur_browser={cur_browser_type}'
+    log.info(f'rerun_cmd_str: {rerun_cmd_str}')
+
+    pool = Pool(processes) if len(features) >= processes else Pool(
+        len(features))
+    results = pool.map(
+        partial(execute_parallel_feature, behave_cmd=rerun_cmd_str,
+                feature_path=rerun_feature_path, browser_type=browser_type),
+        features)
+    pool.close()
+    pool.join()
+    log.info(f'[parallel_rerun] result: {results}')
